+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Search list of quotes by reference name or creator email and filter by cost center name
+
 ## [1.4.0] - 2022-03-10
 
 ### Added